Capitalization and multiple commands support
This commit is addressing issues 1 and 2.
Issue 1 has to do with parsing multiple commands out of the same message. There are times where a player may type "!a !b" and miss the enter key in the middle of those two commands. Currently there is command splitting support, and the commands are split by spacebar. Read github issue for more details.

Issue 2 has to deal with lack of capitalization support. Now all commands are parsed to lowercase immediately. You can type "!LEft" or "!LEFT" and have all commands be parsed correctly.